+++ src/pages/main/dataSlider.tsx
@@ -1,0 +1,26 @@
+import { DataSliderItem } from '../../components/slider/Slider';
+
+const dataSlider: DataSliderItem[] = [
+  {
+    id: 'img1',
+    name: 'main-img-1',
+  },
+  {
+    id: 'img2',
+    name: 'main-img-2',
+  },
+  {
+    id: 'img3',
+    name: 'main-img-3',
+  },
+  {
+    id: 'img4',
+    name: 'main-img-4',
+  },
+  {
+    id: 'img5',
+    name: 'main-img-5',
+  },
+];
+
+export default dataSlider;
